+++ scripts/macros/monsterFeatures/deepGnome/deepGnome.js
@@ -1,0 +1,4 @@
+import {stoneCamouflage} from './stoneCamouflage.js'
+export let deepGnome = {
+    'stoneCamouflage': stoneCamouflage
+}